Trabalhando nos exercícios

+++ aula04-FundamentosPOOExercises/src/Exercicio3/Aluno.java
@@ -5,7 +5,7 @@ public class Aluno {
     public double nota1;
     public double nota2;
 
-    public Aluno(String nome, int nota1, int nota2) {
+    public Aluno(String nome, double nota1, double nota2) {
         this.nome = nome;
         this.nota1 = nota1;
         this.nota2 = nota2;
